feat(diary): added classes for user part of the project
