+++ src/main/java/com/diary/demo/controller/UserController.java
@@ -1,0 +1,4 @@
+package com.diary.demo.controller;
+
+public class UserController {
+}
