+++ src/main/java/com/diary/demo/diaryDto/DiaryCreateErrorResponseDto.java
@@ -1,11 +1,11 @@
-package com.diary.demo.dto;
+package com.diary.demo.diaryDto;
 
-public class DiaryDeleteErrorResponseDto {
+public class DiaryCreateErrorResponseDto {
 
     private final int status;
     private final String message;
 
-    public DiaryDeleteErrorResponseDto(int status, String message) {
+    public DiaryCreateErrorResponseDto(int status, String message) {
         this.status = status;
         this.message = message;
     }
@@ -17,4 +17,7 @@ public class DiaryDeleteErrorResponseDto {
     public String getMessage() {
         return message;
     }
+
+    // This object is read-only.
+    // All the values are set in the constructor at once.
 }
