+++ src/main/java/com/diary/demo/diaryDto/DiaryCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.diary.demo.dto;
+package com.diary.demo.diaryDto;
 
 import org.springframework.web.multipart.MultipartFile;
