+++ src/main/java/com/diary/demo/diaryDto/DiaryDeleteErrorResponseDto.java
@@ -1,10 +1,11 @@
-package com.diary.demo.dto;
+package com.diary.demo.diaryDto;
 
-public class DiaryUpdateErrorResponseDto {
+public class DiaryDeleteErrorResponseDto {
+
     private final int status;
     private final String message;
 
-    public DiaryUpdateErrorResponseDto(int status, String message) {
+    public DiaryDeleteErrorResponseDto(int status, String message) {
         this.status = status;
         this.message = message;
     }
@@ -12,6 +13,7 @@ public class DiaryUpdateErrorResponseDto {
     public int getStatus() {
         return status;
     }
+
     public String getMessage() {
         return message;
     }
