+++ src/main/java/com/diary/demo/diaryDto/DiaryDetailRequestDto.java
@@ -1,4 +1,4 @@
-package com.diary.demo.dto;
+package com.diary.demo.diaryDto;
 
 public class DiaryDetailRequestDto {
 }
