+++ src/main/java/com/diary/demo/diaryDto/DiaryUpdateErrorResponseDto.java
@@ -1,10 +1,10 @@
-package com.diary.demo.dto;
+package com.diary.demo.diaryDto;
 
-public class DiaryDetailErrorResponseDto {
+public class DiaryUpdateErrorResponseDto {
     private final int status;
     private final String message;
 
-    public DiaryDetailErrorResponseDto(int status, String message) {
+    public DiaryUpdateErrorResponseDto(int status, String message) {
         this.status = status;
         this.message = message;
     }
