+++ src/main/java/com/diary/demo/diaryDto/DiaryUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.diary.demo.dto;
+package com.diary.demo.diaryDto;
 
 public class DiaryUpdateRequestDto {
     private String title;
