+++ src/main/java/com/diary/demo/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package com.diary.demo.repository;
+
+import com.diary.demo.domain.Users;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<Users, Long> {
+}
