+++ src/main/java/com/diary/demo/service/UserService.java
@@ -1,0 +1,4 @@
+package com.diary.demo.service;
+
+public class UserService {
+}
