+++ src/main/java/com/diary/demo/userDto/UserCreateErrorResponseDto.java
@@ -1,0 +1,4 @@
+package com.diary.demo.userDto;
+
+public class UserCreateErrorResponseDto {
+}
