+++ src/main/java/com/diary/demo/userDto/UserCreateRequestDto.java
@@ -1,0 +1,4 @@
+package com.diary.demo.userDto;
+
+public class UserCreateRequestDto {
+}
